backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

SQLAlchemy==2.0.27
psycopg[binary]==3.1.18

pydantic==2.6.1
pydantic-settings==2.2.1

python-jose==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

requests==2.31.0

celery==5.3.6
redis==5.0.1

alembic==1.11.1
python-dotenv==1.0.0
httpx==0.24.1
gunicorn==20.1.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.10.0
factory-boy==3.2.1
Faker==18.9.0
email-validator==2.0.0
itsdangerous==2.1.2
aiofiles==23.1.0
Jinja2==3.1.4
python-multipart==0.0.6

cryptography==42.0.8
celery[redis]==5.3.6


openpyxl==3.1.2
argon2-cffi==23.1.0
